refact(brain-global-save-button): uses store to centralize Brain state
